+++ hypervisor/include/arch/x86/guest/guest.h
@@ -58,6 +58,9 @@ enum {
 
 struct vhm_request;
 
+int acrn_insert_request_wait(struct vcpu *vcpu, struct vhm_request *req);
+int get_req_info(char *str, int str_max);
+
 /*
  * VCPU related APIs
  */
